pyproject.toml
[project]
name = "fastauth"
version = "0.1.0"
description = "High-performance authentication system with FastAPI and async PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "sqlmodel>=0.0.24",
    "psycopg[asyncpg]>=3.2.9",
    "asyncpg>=0.30.0",
    "alembic>=1.16.4",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastauth"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
